add tests for text replacement
