+++ sudachi/src/input_text/input_buffer/edit.rs
@@ -19,7 +19,6 @@ use std::ops::Range;
 pub struct ReplaceOp<'a> {
     what: Range<usize>,
     with: ReplaceTgt<'a>,
-    nchars: usize,
 }
 
 enum ReplaceTgt<'a> {
@@ -45,7 +44,6 @@ impl<'a> EditInput<'a> {
             // This reference can be created ONLY inside with_replacer call.
             // This reference lifetime will end before returning from with_replacer call.
             with: ReplaceTgt::Ref(unsafe { std::mem::transmute(result) }),
-            nchars: result.chars().count(),
         };
         self.replaces.push(op);
     }
@@ -55,7 +53,6 @@ impl<'a> EditInput<'a> {
         let op = ReplaceOp {
             what: range,
             with: ReplaceTgt::Char(result),
-            nchars: 1,
         };
         self.replaces.push(op);
     }
@@ -66,7 +63,6 @@ impl<'a> EditInput<'a> {
         let op = ReplaceOp {
             what: range,
             with: ReplaceTgt::Str(result),
-            nchars,
         };
         self.replaces.push(op);
     }
@@ -97,29 +93,16 @@ pub fn resolve_edits(
     edits: &mut Vec<ReplaceOp>,
 ) {
     let mut start: usize = 0;
-    let mut offset: isize = 0;
     for edit in edits.drain(..) {
         target.push_str(&source[start..edit.what.start]);
-        target_mapping.extend(
-            source_mapping[start..edit.what.start]
-                .iter()
-                .map(|x| (*x as isize + offset) as usize),
-        );
+        target_mapping.extend(source_mapping[start..edit.what.start].iter());
         start = edit.what.end;
-        offset += match edit.with {
-            ReplaceTgt::Str(s) => add_replace(
-                offset,
-                source_mapping,
-                target,
-                target_mapping,
-                edit.what,
-                &s,
-            ),
-            ReplaceTgt::Ref(s) => {
-                add_replace(offset, source_mapping, target, target_mapping, edit.what, s)
+        match edit.with {
+            ReplaceTgt::Str(s) => {
+                add_replace(source_mapping, target, target_mapping, edit.what, &s)
             }
+            ReplaceTgt::Ref(s) => add_replace(source_mapping, target, target_mapping, edit.what, s),
             ReplaceTgt::Char(c) => add_replace(
-                offset,
                 source_mapping,
                 target,
                 target_mapping,
@@ -129,48 +112,40 @@ pub fn resolve_edits(
         }
     }
     target.push_str(&source[start..]);
-    target_mapping.extend(
-        source_mapping[start..]
-            .iter()
-            .map(|x| (*x as isize + offset) as usize),
-    );
+    target_mapping.extend(source_mapping[start..].iter());
 }
 
 fn add_replace(
-    offset: isize,
     source_mapping: &Vec<usize>,
     target: &mut String,
     target_mapping: &mut Vec<usize>,
     what: Range<usize>,
     with: &str,
-) -> isize {
-    // 0 1 2 3 [4 5] 6 7
-
+) {
     target.push_str(with);
     let old_mapping = &source_mapping[what.clone()];
-    let old_len = what.len() as isize;
-    let new_len = with.len() as isize;
+    let old_len = what.len();
+    let new_len = with.len();
     if new_len >= old_len {
         for i in 0..old_len {
-            target_mapping.push((old_mapping[i as usize] as isize + offset) as usize);
+            target_mapping.push(old_mapping[i]);
         }
-        let last_value = *old_mapping.last().unwrap_or(&0) as isize + offset;
+        let last_value = source_mapping[what.end];
         for _ in old_len..new_len {
-            target_mapping.push(last_value as usize);
+            target_mapping.push(last_value);
         }
     } else {
         for i in 0..new_len {
-            target_mapping.push((old_mapping[i as usize] as isize + offset) as usize);
+            target_mapping.push(old_mapping[i]);
         }
     }
-
-    new_len - old_len
 }
 
 #[cfg(test)]
 mod test {
     use super::super::InputBuffer;
     use super::*;
+    use crate::input_text::InputTextIndex;
 
     #[test]
     fn edit_ref_1() {
@@ -182,6 +157,118 @@ mod test {
             })
             .expect("should not break");
         assert_eq!(0, buffer.replaces.len());
-        assert_eq!(buffer.current(), "宇銀人")
+        assert_eq!(buffer.current(), "宇銀人");
+        assert_eq!(buffer.orig_slice(3..6), "宙");
+    }
+
+    #[test]
+    fn edit_char_1() {
+        let mut buffer = InputBuffer::from("宇宙人");
+        buffer
+            .with_replacer(|_, mut r| {
+                r.replace_char(3..6, '銀');
+                Ok(r)
+            })
+            .expect("should not break");
+        assert_eq!(0, buffer.replaces.len());
+        assert_eq!(buffer.current(), "宇銀人");
+        assert_eq!(buffer.orig_slice(3..6), "宙");
+    }
+
+    #[test]
+    fn edit_str_1() {
+        let mut buffer = InputBuffer::from("宇宙人");
+        buffer
+            .with_replacer(|_, mut r| {
+                r.replace_own(3..6, String::from("銀"));
+                Ok(r)
+            })
+            .expect("should not break");
+        assert_eq!(0, buffer.replaces.len());
+        assert_eq!(buffer.current(), "宇銀人");
+        assert_eq!(buffer.orig_slice(3..6), "宙");
+    }
+
+    #[test]
+    fn replace_start_w_longer() {
+        let mut buffer = InputBuffer::from("宇宙人");
+        buffer
+            .with_replacer(|_, mut r| {
+                r.replace_ref(0..3, "銀河");
+                Ok(r)
+            })
+            .expect("should not break");
+        assert_eq!(0, buffer.replaces.len());
+        assert_eq!(buffer.m2o, &[0, 1, 2, 3, 3, 3, 3, 4, 5, 6, 7, 8, 9]);
+        assert_eq!(buffer.current(), "銀河宙人");
+        assert_eq!(buffer.orig_slice(0..6), "宇");
+        assert_eq!(buffer.orig_slice(0..3), "宇");
+        assert_eq!(buffer.orig_slice(3..6), "");
+    }
+
+    #[test]
+    fn replace_mid_w_longer() {
+        let mut buffer = InputBuffer::from("宇宙人");
+        buffer
+            .with_replacer(|_, mut r| {
+                r.replace_ref(3..6, "銀河");
+                Ok(r)
+            })
+            .expect("should not break");
+        assert_eq!(0, buffer.replaces.len());
+        assert_eq!(buffer.m2o, &[0, 1, 2, 3, 4, 5, 6, 6, 6, 6, 7, 8, 9]);
+        assert_eq!(buffer.current(), "宇銀河人");
+        assert_eq!(buffer.orig_slice(3..9), "宙");
+        assert_eq!(buffer.orig_slice(3..6), "宙");
+        assert_eq!(buffer.orig_slice(6..9), "");
+    }
+
+    #[test]
+    fn replace_end_w_longer() {
+        let mut buffer = InputBuffer::from("宇宙人");
+        buffer
+            .with_replacer(|_, mut r| {
+                r.replace_ref(6..9, "銀河");
+                Ok(r)
+            })
+            .expect("should not break");
+        assert_eq!(0, buffer.replaces.len());
+        assert_eq!(buffer.m2o, &[0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 9, 9, 9]);
+        assert_eq!(buffer.current(), "宇宙銀河");
+        assert_eq!(buffer.orig_slice(6..12), "人");
+        assert_eq!(buffer.orig_slice(6..9), "人");
+        assert_eq!(buffer.orig_slice(9..12), "");
+    }
+
+    #[test]
+    fn replace_end_w_shorter() {
+        let mut buffer = InputBuffer::from("宇宙人");
+        buffer
+            .with_replacer(|_, mut r| {
+                r.replace_ref(3..9, "河");
+                Ok(r)
+            })
+            .expect("should not break");
+        assert_eq!(0, buffer.replaces.len());
+        assert_eq!(buffer.current(), "宇河");
+        assert_eq!(buffer.m2o, &[0, 1, 2, 3, 4, 5, 9]);
+        assert_eq!(buffer.orig_slice(0..3), "宇");
+        assert_eq!(buffer.orig_slice(3..6), "宙人");
+    }
+
+    #[test]
+    fn replace_start_w_shorter() {
+        let mut buffer = InputBuffer::from("宇宙人");
+        buffer
+            .with_replacer(|_, mut r| {
+                r.replace_ref(0..6, "河");
+                Ok(r)
+            })
+            .expect("should not break");
+        assert_eq!(0, buffer.replaces.len());
+        assert_eq!(buffer.current(), "河人");
+        assert_eq!(buffer.m2o, &[0, 1, 2, 6, 7, 8, 9]);
+        assert_eq!(buffer.orig_slice(0..3), "宇宙");
+        assert_eq!(buffer.orig_slice(3..6), "人");
     }
 }
