+++ sudachi/src/input_text/input_buffer/mod.rs
@@ -23,8 +23,10 @@ use crate::dic::category_type::CategoryType;
 use crate::dic::grammar::Grammar;
 use std::ops::Range;
 
-use crate::error::SudachiResult;
-use crate::input_text::PathRewriteAPI;
+use crate::error::{SudachiError, SudachiResult};
+use crate::input_text::InputTextIndex;
+
+const MAX_LENGTH: usize = 32 * 1024;
 
 #[derive(Eq, PartialEq, Debug)]
 enum BufferState {
@@ -41,17 +43,34 @@ impl Default for BufferState {
 
 #[derive(Default)]
 pub struct InputBuffer {
+    /// Original input data, output is done on this
     original: String,
+    /// Normalized input data, analysis is done on this. Byte-based indexing.
     modified: String,
-    modified2: String,
-    m2o_bytes: Vec<usize>,
-    m2o_bytes2: Vec<usize>,
+    /// Buffer for normalization, reusing allocations
+    modified_2: String,
+    /// Byte mapping from normalized data to originals.
+    /// Only values lying on codepoint boundaries are correct. Byte-based indexing.
+    m2o: Vec<usize>,
+    /// Buffer for normalization
+    m2o_2: Vec<usize>,
+    /// Characters of the modified string. Char-based indexing.
     mod_chars: Vec<char>,
+    /// Char-to-byte mapping for the modified string. Char-based indexing.
     mod_c2b: Vec<usize>,
+    /// Byte-to-char mapping for the modified string. Byte-based indexing.
     mod_b2c: Vec<usize>,
+    /// Markers whether the byte can start new word or not
     mod_bow: Vec<bool>,
+    /// Character categories. Char-based indexing.
+    mod_cat: Vec<CategoryType>,
+    /// Number of codepoints with the same category. Char-based indexing.
     mod_category_continuity: Vec<usize>,
+    /// This very temporaliy keeps the replace data.
+    /// 'static lifetime is a lie, it is incorrect to use it outside `with_replacer` function
+    /// or its callees.
     replaces: Vec<edit::ReplaceOp<'static>>,
+    /// Current state of the buffer
     state: BufferState,
 }
 
@@ -60,10 +79,12 @@ impl InputBuffer {
         InputBuffer::default()
     }
 
+    /// Resets the input buffer, so it could be used to process new input.
+    /// New input should be written to the returned mutable reference.
     pub fn reset(&mut self) -> &mut String {
         self.original.clear();
         self.modified.clear();
-        self.m2o_bytes.clear();
+        self.m2o.clear();
         self.mod_chars.clear();
         self.mod_c2b.clear();
         self.mod_bow.clear();
@@ -72,18 +93,25 @@ impl InputBuffer {
         &mut self.original
     }
 
+    /// Creates input from the passed string. Should be used mostly for tests.
+    ///
+    /// Panics if the input string is too long.
     pub fn from<'a, T: AsRef<str>>(data: T) -> InputBuffer {
         let mut buf = Self::new();
         buf.reset().push_str(data.as_ref());
-        buf.start_build();
+        buf.start_build().expect("");
         buf
     }
 
-    pub fn start_build(&mut self) {
+    pub fn start_build(&mut self) -> SudachiResult<()> {
+        if self.original.len() > MAX_LENGTH {
+            return Err(SudachiError::InputTooLong(self.original.len(), MAX_LENGTH));
+        }
         debug_assert_eq!(self.state, BufferState::Clean);
         self.state = BufferState::RW;
         self.modified.push_str(&self.original);
-        self.m2o_bytes.extend(0..self.modified.len());
+        self.m2o.extend(0..self.modified.len() + 1);
+        Ok(())
     }
 
     pub fn build(&mut self, _grammar: &Grammar) -> SudachiResult<()> {
@@ -106,13 +134,13 @@ impl InputBuffer {
         }
         edit::resolve_edits(
             &self.modified,
-            &self.m2o_bytes,
-            &mut self.modified2,
-            &mut self.m2o_bytes2,
+            &self.m2o,
+            &mut self.modified_2,
+            &mut self.m2o_2,
             &mut self.replaces,
         );
-        std::mem::swap(&mut self.modified, &mut self.modified2);
-        std::mem::swap(&mut self.m2o_bytes, &mut self.m2o_bytes2);
+        std::mem::swap(&mut self.modified, &mut self.modified_2);
+        std::mem::swap(&mut self.m2o, &mut self.m2o_2);
     }
 
     fn rollback(&mut self) {
@@ -144,6 +172,7 @@ impl InputBuffer {
         }
     }
 
+    /// Recompute chars from modified string (useful if the processing will use chars)
     pub fn refresh_chars(&mut self) {
         debug_assert_eq!(self.state, BufferState::RW);
         if self.mod_chars.is_empty() {
@@ -177,30 +206,57 @@ impl InputBuffer {
     }
 
     pub fn can_bow(&self, offset: usize) -> bool {
-        todo!()
+        debug_assert_eq!(self.state, BufferState::RO);
+        let cpidx = self.mod_b2c[offset];
+        self.mod_bow[cpidx]
     }
 
     pub fn get_word_candidate_length(&self, offset: usize) -> usize {
+        debug_assert_eq!(self.state, BufferState::RO);
         todo!()
     }
 
-    pub fn orig_slice(&self, range: Range<usize>) -> &str {
-        todo!()
+    pub fn to_orig(&self, range: Range<usize>) -> Range<usize> {
+        debug_assert_ne!(self.state, BufferState::Clean);
+        self.m2o[range.start]..self.m2o[range.end]
     }
 }
 
-impl PathRewriteAPI for InputBuffer {
+impl InputTextIndex for InputBuffer {
     fn cat_of_range(&self, range: Range<usize>) -> CategoryType {
         todo!()
     }
 
     fn cat_at_byte(&self, offset: usize) -> CategoryType {
-        todo!()
+        debug_assert_eq!(self.state, BufferState::RO);
+        let cpidx = self.mod_b2c[offset];
+        self.mod_cat[cpidx]
     }
 
     fn num_codepts(&self, range: Range<usize>) -> usize {
+        debug_assert_eq!(self.state, BufferState::RO);
         let start = self.mod_b2c[range.start];
         let end = self.mod_b2c[range.end];
-        end - start
+        end - start - 1
+    }
+
+    fn cat_continuous_len(&self, offset: usize) -> usize {
+        debug_assert_eq!(self.state, BufferState::RO);
+        todo!()
+    }
+
+    fn byte_distance(&self, byte: usize, codepts: usize) -> usize {
+        debug_assert_eq!(self.state, BufferState::RO);
+        todo!()
+    }
+
+    fn orig_slice(&self, range: Range<usize>) -> &str {
+        debug_assert_ne!(self.state, BufferState::Clean);
+        &self.original[self.to_orig(range)]
+    }
+
+    fn curr_slice(&self, range: Range<usize>) -> &str {
+        debug_assert_ne!(self.state, BufferState::Clean);
+        &self.modified[range]
     }
 }
