+++ sudachi/src/plugin/oov/simple_oov/mod.rs
@@ -22,7 +22,7 @@ use crate::config::Config;
 use crate::dic::grammar::Grammar;
 use crate::dic::lexicon::word_infos::WordInfo;
 use crate::input_text::input_buffer::InputBuffer;
-use crate::input_text::Utf8InputText;
+use crate::input_text::{InputTextIndex, Utf8InputText};
 use crate::plugin::oov::OovProviderPlugin;
 use crate::prelude::*;
